+++ client/views/__init__.py
@@ -7,6 +7,7 @@ from .user_document import user_document
 from .user_document_upload import user_document_upload
 from .apis import *
 from .users import users
-from .user_perms import user_perms, user_set_perms
+from .user_perms import user_perms
+from .user_set_perms import user_set_perms
 from .create_new_user import create_new_user
 from .user_delete import user_delete
