+++ templates/educational/uni_show_docs.html
@@ -7,7 +7,13 @@
 
 {% endblock %}
 {% block js %}
-
+    <script>
+        $(window).on('load', function () {
+            setTimeout(function () {
+                $('#message').css('display', 'none');
+            }, 3000)
+        })
+    </script>
 {% endblock %}
 {% block bodyclass %} class="nk-body bg-lighter has-rtl npc-default pg-auth"{% endblock %}
 {% block body %}
@@ -36,7 +42,7 @@
                                     <div class="card">
                                         <div class="card card-inner">
                                             {% if message %}
-                                                <div class="example-alert mb-3">
+                                                <div class="example-alert mb-3" id="message">
                                                     <div class="alert alert-success alert-icon">
                                                         <em class="icon ni ni-check-circle"></em>
                                                         تغییرات ذخیره شد
@@ -48,13 +54,13 @@
                                                       method="post">
                                                     {% csrf_token %}
                                                     <div class="row gy-4">
-                                                        <div class="col-sm-6">
+                                                        <div class="col-sm-4">
                                                             <div class="form-group">
                                                                 <label class="form-label text text-indigo"
                                                                        for="doc">مدارک موردنیاز دانشگاه
                                                                 </label>
                                                                 <div class="form-control-wrap">
-                                                                <textarea
+                                                                    <textarea
                                                                         class="form-control no-resize"
                                                                         id="doc"
                                                                         placeholder="کارت سربازی
@@ -64,13 +70,13 @@
                                                                 </div>
                                                             </div>
                                                         </div>
-                                                        <div class="col-sm-6">
+                                                        <div class="col-sm-8">
                                                             <div class="form-group">
                                                                 <label class="form-label text text-indigo"
                                                                        for="comment">توضیحات دانشگاه
                                                                 </label>
                                                                 <div class="form-control-wrap">
-                                                                <textarea
+                                                                    <textarea
                                                                         class="form-control no-resize"
                                                                         id="comment"
                                                                         placeholder="مورد اول
